feat(GPACalculator): Refactor and enhance GPA calculator functionality and UI - Introduced a new GPACalculator component to encapsulate all GPA-related features, including course management and GPA calculation - Updated CourseList to handle course removal and reset functionality more intuitively - Improved UI elements with enhanced styling and animations for a better user experience - Streamlined the GPA calculation logic and integrated animated UI elements for visual appeal
