+++ src/components/tools/GPACalculator/CourseList.tsx
@@ -83,12 +83,18 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
   };
 
   /**
-   * Removes a course from the list if there's more than one course
+   * Removes a course from the list or resets it if it's the last one
    * @param id - ID of the course to remove
    */
   const removeCourse = (id: number) => {
     if (courses.length > 1) {
+      // If there are multiple courses, remove the selected one
       onCoursesChange(courses.filter(course => course.id !== id));
+    } else {
+      // If it's the last course, reset its values instead of removing
+      onCoursesChange([
+        { id: courses[0].id, name: '', marks: '', credits: 0, grade: '-', gradePoints: 0 }
+      ]);
     }
   };
 
@@ -112,7 +118,7 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
               type="text"
               value={course.name}
               onChange={(e) => handleNameChange(course.id, e.target.value)}
-              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm"
+              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm focus:ring-2 focus:ring-indigo-400 focus:border-indigo-400 transition-all duration-200"
               placeholder="Enter course name"
             />
           </div>
@@ -123,7 +129,7 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
               type="number"
               value={course.credits || ''}
               onChange={(e) => handleCreditsChange(course.id, e.target.value)}
-              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm"
+              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm focus:ring-2 focus:ring-indigo-400 focus:border-indigo-400 transition-all duration-200"
               placeholder="Credits"
               min="0"
             />
@@ -135,7 +141,7 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
               type="number"
               value={course.marks || ''}
               onChange={(e) => handleMarksChange(course.id, e.target.value)}
-              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm"
+              className="w-full rounded-lg border-gray-200 p-2 text-sm shadow-sm focus:ring-2 focus:ring-indigo-400 focus:border-indigo-400 transition-all duration-200"
               placeholder="Marks"
               min="0"
               max="100"
@@ -144,17 +150,19 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
 
           {/* Grade Display */}
           <div className="col-span-6 sm:col-span-2">
-            <span className="inline-flex items-center justify-center px-3 py-2 text-sm font-medium text-indigo-600 bg-indigo-100 rounded-lg">
+            <span className={`inline-flex items-center justify-center px-3 py-2 text-sm font-medium rounded-lg w-full
+              ${course.grade === 'F' ? 'bg-red-100 text-red-800' : 
+                course.grade === '-' ? 'bg-gray-100 text-gray-600' : 
+                'bg-green-100 text-green-800'}`}>
               {course.grade}
             </span>
           </div>
 
           {/* Remove Course Button */}
-          <div className="col-span-6 sm:col-span-2">
+          <div className="col-span-6 sm:col-span-2 flex justify-center">
             <button
               onClick={() => removeCourse(course.id)}
-              className="text-red-600 hover:text-red-700 transition-colors"
-              disabled={courses.length === 1}
+              className="text-red-500 hover:text-red-700 transition-colors p-2 rounded-lg hover:bg-red-50"
               aria-label="Remove course"
             >
               <TrashIcon className="w-5 h-5" />
@@ -166,7 +174,7 @@ const CourseList: React.FC<CourseListProps> = ({ courses, isTU, onCoursesChange 
       {/* Add Course Button */}
       <button
         onClick={addCourse}
-        className="flex items-center gap-2 text-indigo-600 hover:text-indigo-700 transition-colors"
+        className="flex items-center gap-2 text-indigo-600 hover:text-indigo-700 transition-colors mt-4 px-4 py-2 rounded-lg hover:bg-indigo-50"
         aria-label="Add new course"
       >
         <PlusIcon className="w-5 h-5" />
