+++ src/components/tools/GPACalculator/index.ts
@@ -2,11 +2,13 @@
  * GPA Calculator Module
  * 
  * This barrel file exports all components and utilities needed for the GPA Calculator feature.
- * - Components: CourseList and UniversityToggle
+ * - Main Component: GPACalculator
+ * - Sub-components: CourseList and UniversityToggle
  * - Types: Course, GradeScaleItem, and University types
  * - Utilities: Grade calculation functions and grade scales
  */
 
+export { default } from './GPACalculator';
 export { default as CourseList } from './CourseList';
 export { default as UniversityToggle } from './UniversityToggle';
 export * from './types';
